Clean up the implementation of roots() and move it to Graphs. This is part of the effort to delete GraphProperties.java, as discussed earlier. The other method in GraphProperties is isCyclic(), which I'd like to handle separately (e.g. it should handle undirected graphs if it's going to be in Graphs, or maybe moved to labs/). ------------- Created by MOE: https://github.com/google/moe MOE_MIGRATED_REVID=125176121
